Arreglos

+++ Foro_UAL/src/main/java/com/MDS2/ForoUal/Interfaz/Visualizar_Notificaciones.java
@@ -6,7 +6,7 @@ import com.MDS2.ForoUal.Backend.ORM.src.Notificacion;
 public class Visualizar_Notificaciones extends Visualizar_Notificaciones_Ventana{
 
 	public Visualizar_Notificaciones () {
-		Notificacion [] n = foroUI.db.Cargar_Notificaciones((int)foroUI.user.getORMID());
+		Notificacion [] n = foroUI.db.Cargar_Notificaciones(foroUI.user.getORMID());
 		System.out.println("longitud " +n.length);
 		
 		for(int i = 0; i < Math.min(n.length, 7); i++)
